perf: uint8 E8M0 scale storage + in-kernel decode + fused reduction
Store MXFP4 expert scales as raw uint8 (1 byte, E8M0 exponent bytes)
instead of float32 (4 bytes). Decode in Triton kernel via exp2(x - 127)
—a single hardware instruction on Blackwell—eliminating all host-side
dtype conversion overhead.

Additional kernel optimizations:
- Fused dot product: merge two tl.sum() calls into one, reducing
  register pressure and synchronization points per K-tile iteration
- Expanded autotune configs: add num_stages=3 variants for SM120's
  99 KB SMEM, more warp configs for larger tiles
- FlashMLA: add num_stages=3 and BLOCK_T=8 autotune options

Result on SM120 (2× RTX PRO 6000, TP=2, EAGLE 1/1/2):
- Decode: 44.5 → 47.5 tok/s (+6.7%)
- Token pool: 117K → 397K (+240%)
- Scale memory: ~12.7 GB/GPU (FP32) → ~3.2 GB/GPU (uint8)

+++ python/sglang/srt/layers/quantization/fp8.py
@@ -998,14 +998,15 @@ class Fp8MoEMethod(FusedMoEMethodBase):
         # WEIGHT_SCALES
         if self.is_fp4_expert:
             fp4_block_k = 32
-            # Use bfloat16 to save ~8.5GB/GPU vs FP32.
-            # Simple upcast to FP32 at runtime has negligible overhead.
+            # Use float32 for correct weight loading (checkpoint stores e8m0
+            # which numerically converts to float32).  process_weights_after_loading
+            # converts to uint8 (1 byte) for SM120 Triton in-kernel decode.
             w13_weight_scale = torch.nn.Parameter(
                 torch.ones(
                     num_experts,
                     2 * intermediate_size_per_partition,
                     hidden_size // fp4_block_k,
-                    dtype=torch.bfloat16,
+                    dtype=torch.float32,
                 ),
                 requires_grad=False,
             )
@@ -1014,7 +1015,7 @@ class Fp8MoEMethod(FusedMoEMethodBase):
                     num_experts,
                     hidden_size,
                     intermediate_size_per_partition // fp4_block_k,
-                    dtype=torch.bfloat16,
+                    dtype=torch.float32,
                 ),
                 requires_grad=False,
             )
